% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{amort_cd}
\alias{amort_cd}
\title{Constant dollar amortization method}
\usage{
amort_cd(p, i, m, end = FALSE, skipY1 = FALSE)
}
\arguments{
\item{p}{Principle}

\item{i}{interest rate}

\item{m}{Amortization period}

\item{end}{Logical. If TRUE, payments are made at the end of periods}

\item{skipY1}{Logial. If TRUE, year-1 payment is 0}
}
\value{
A vector of annual amortization payments
}
\description{
Constant dollar amortization method
}
