% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{amort_cp}
\alias{amort_cp}
\title{Constant percent amortization method}
\usage{
amort_cp(p, i, m, g, end = FALSE, skipY1 = FALSE)
}
\arguments{
\item{p}{Principle}

\item{i}{interest rate}

\item{m}{Amortization period}

\item{g}{Constant growth rate of payments}

\item{end}{Logical. If TRUE, payments are made at the end of periods}

\item{skipY1}{Logial. If TRUE, year-1 payment is 0}
}
\value{
A vector of annual amortization payments
}
\description{
Constant percent amortization method
}
