% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{amort_LG}
\alias{amort_LG}
\title{Function for choosing amortization methods}
\usage{
amort_LG(p, i, m, g, end = FALSE, method = "cd", skipY1 = FALSE)
}
\arguments{
\item{p}{Principle}

\item{i}{interest rate}

\item{m}{Amortization period}

\item{g}{Constant growth rate of payments}

\item{end}{Logical. If TRUE, payments are made at the end of periods}

\item{method}{Amortization method}

\item{skipY1}{Logial. If TRUE, year-1 payment is 0}
}
\value{
A vector of annual amortization payments
}
\description{
Function for choosing amortization methods
}
