% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{amort_sl}
\alias{amort_sl}
\title{Strait line method}
\usage{
amort_sl(p, i, m, end = FALSE)
}
\arguments{
\item{p}{Principle}

\item{i}{interest rate}

\item{m}{Amortization period}

\item{end}{Logical. If TRUE, payments are made at the end of periods}
}
\value{
A vector of annual amortization payments
}
\description{
Strait line method
}
