% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fnC}
\alias{fnC}
\title{Account balances}
\usage{
fnC(balance, C, r)
}
\arguments{
\item{balance}{vector to store balances}

\item{C}{Cash flow}

\item{r}{Returns}
}
\value{
Balances
}
\description{
Account balances
}
