% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gaip}
\alias{gaip}
\title{Amortization function with constant-growth payments at each period}
\usage{
gaip(p, i, n, g, end = FALSE)
}
\arguments{
\item{p}{Principal}

\item{i}{interest rate}

\item{n}{Amortization period}

\item{g}{Constant growth rate of payments}

\item{end}{Logical. If TRUE, payments are made at the end of periods}
}
\value{
A vector of annual amortization payments
}
\description{
Amortization function with constant-growth payments at each period
}
