% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_geoReturn}
\alias{get_geoReturn}
\title{Calculating geometric mean return}
\usage{
get_geoReturn(x)
}
\arguments{
\item{x}{Vector of returns}
}
\value{
Geometric mean return
}
\description{
Calculating geometric mean return
}
