% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_PVB_active}
\alias{get_PVB_active}
\title{Function for calculating PVB for actives}
\usage{
get_PVB_active(
  age_fn,
  ea_fn,
  age_ret,
  benefit_init,
  cola_assumed,
  i,
  decrement = df_decrement,
  age_max_ = age_max
)
}
\arguments{
\item{age_fn}{Current age}

\item{ea_fn}{entry age}

\item{age_ret}{Retirement age}

\item{benefit_init}{Benefit payment at retirement age (age_ret)}

\item{cola_assumed}{Assumed COLA}

\item{i}{Discount rate}

\item{decrement}{Decrement table to use}

\item{age_max_}{Maximum age}
}
\value{
Actuarial present value of active members
}
\description{
Function for calculating PVB for actives
}
