% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_PVB_retiree}
\alias{get_PVB_retiree}
\title{Calculates actuarial PV for given age, initial benefit,
cola assumption, and mortality}
\usage{
get_PVB_retiree(
  age_fn,
  benefit_init,
  cola_assumed,
  i,
  decrement = df_decrement,
  age_max_ = age_max
)
}
\arguments{
\item{age_fn}{Current age}

\item{benefit_init}{Current benefit payment}

\item{cola_assumed}{Assumed future annual cola}

\item{i}{Discount rate}

\item{decrement}{Decrement table to use}

\item{age_max_}{Maximum age}
}
\value{
Actuarial present value of future benefits
}
\description{
Calculates actuarial PV for given age, initial benefit,
cola assumption, and mortality
}
