% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{na2zero}
\alias{na2zero}
\title{Turning NAs and NaNs to 0}
\usage{
na2zero(x)
}
\arguments{
\item{x}{A vector that may contain NAs or NaNs}
}
\value{
The vector with NAs and NaNs replaced by 0
}
\description{
Turning NAs and NaNs to 0
}
