% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pmt}
\alias{pmt}
\title{Amortization function with constant payment amounts at each period}
\usage{
pmt(p, i, n, end = FALSE)
}
\arguments{
\item{p}{principle}

\item{i}{interest rate}

\item{n}{amortization period}

\item{end}{Logical. If TRUE, payments are made at the end of periods}
}
\value{
A vector of annual amortization payments
}
\description{
Amortization function with constant payment amounts at each period
}
